Move authentication-related functions from wagtail.admin.decorators to wagtail.admin.auth
